+++ packages/mongodb-access-adapter/datagrip-access-adapter/src/main/kotlin/com/mongodb/jbplugin/accessadapter/datagrip/DataGripBasedReadModelProvider.kt
@@ -11,12 +11,14 @@ import com.intellij.openapi.project.Project
 import com.intellij.psi.util.CachedValue
 import com.intellij.psi.util.CachedValueProvider
 import com.intellij.psi.util.CachedValuesManager
+import com.mongodb.jbplugin.accessadapter.MongoDbDriver
 import com.mongodb.jbplugin.accessadapter.MongoDbReadModelProvider
 import com.mongodb.jbplugin.accessadapter.Slice
 import com.mongodb.jbplugin.accessadapter.datagrip.adapter.DataGripMongoDbDriver
 import kotlinx.coroutines.runBlocking
 
 private typealias MapOfCachedValues = MutableMap<String, CachedValue<*>>
+private typealias DriverFactory = (Project, LocalDataSource) -> MongoDbDriver
 
 /**
  * The service to be injected to access MongoDB. Usually you will use
@@ -37,21 +39,28 @@ private typealias MapOfCachedValues = MutableMap<String, CachedValue<*>>
 class DataGripBasedReadModelProvider(
     private val project: Project,
 ) : MongoDbReadModelProvider<LocalDataSource> {
+    var driverFactory: DriverFactory = { project, dataSource ->
+        DataGripMongoDbDriver(project, dataSource)
+    }
     private val cachedValues: MapOfCachedValues = mutableMapOf()
 
-    override fun <T : Any> slice(dataSource: LocalDataSource, slice: Slice<T>): T = cachedValues
-        .computeIfAbsent(slice.javaClass.canonicalName, fromSlice(dataSource, slice))
-        .value as T
+    override fun <T : Any> slice(
+        dataSource: LocalDataSource,
+        slice: Slice<T>,
+    ): T =
+        cachedValues
+            .computeIfAbsent(slice.javaClass.canonicalName, fromSlice(dataSource, slice))
+            .value as T
 
     private fun <T : Any> fromSlice(
         dataSource: LocalDataSource,
-        slice: Slice<T>
+        slice: Slice<T>,
     ): (String) -> CachedValue<T> {
         val cacheManager = CachedValuesManager.getManager(project)
         return {
             cacheManager.createCachedValue {
                 runBlocking {
-                    val driver = DataGripMongoDbDriver(project, dataSource)
+                    val driver = driverFactory(project, dataSource)
                     val sliceData = slice.queryUsingDriver(driver)
 
                     CachedValueProvider.Result.create(sliceData, dataSource)